Move added modules into tokio
